Split ScanRunner by adding VTRunner.
The ScanRunner takes care of managing hosts etc, while the VTRunner
is only responsible for scanning a given (host, vt) tuple. This is done
to hopefully clean up the separation of responsibilities because the
VTRunner will soon become async, making the logic more complex.
